Square coverage integrated and a test has been added (cant see if it is working)

+++ Point/Point.cpp
@@ -1,4 +1,4 @@
 #include "Point.h"
 Point::Point() {}
 
-Point::Point(signed int x, signed int y): x(x), y(y){}
+Point::Point(double x, double y): x(x), y(y){}
