+++ Point/Point.h
@@ -2,10 +2,10 @@
 
 class Point {
 public:
-	signed int x = 0;
-	signed int y = 0;
+    double x = 0;
+    double y = 0;
     bool visited = false;
 	Point();
-        Point(signed int x, signed int y);
+        Point(double x, double y);
 };
